+++ lib/grizzly_series.dart
@@ -4,8 +4,10 @@
 /// Support for doing something awesome.
 ///
 /// More dartdocs go here.
-library tandas;
+library grizzly_series;
 
+export 'src/array/array.dart';
+export 'src/ndarray/ndarray.dart';
 export 'src/data_frame/data_frame.dart';
 export 'src/series/series.dart';
 export 'src/loaders/loaders.dart';