+++ test/series/int/math_ops.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, SERAGUD. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-import 'package:grizzly/grizzly.dart';
+import 'package:grizzly_series/grizzly_series.dart';
 import 'package:test/test.dart';
 
 void main() {