+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.0.1
 
-- Half functional `DataFrame` and `Series`
+- Half functional `DataFrame`, `Series`, `Array`